+++ Sources/Kommunicate/Classes/Utilities/KmConfigurationSetter.swift
@@ -122,6 +122,14 @@ public class KMConfigurationSetter {
             if let enableDeleteConversationOnLongPress = settingDict["enableDeleteConversationOnLongpress"] as? Bool {
                 Kommunicate.defaultConfiguration.enableDeleteConversationOnLongpress = enableDeleteConversationOnLongPress
             }
+           
+            if let startNewConversationButtonBackgroundColor = settingDict["startNewConversationButtonBackgroundColor"] as? String {
+                Kommunicate.kmConversationViewConfiguration.startNewConversationButtonBackgroundColor = UIColor(hexString: startNewConversationButtonBackgroundColor)
+            }
+            
+            if let startNewConversationButtonTextColor = settingDict["startNewConversationButtonTextColor"] as? String {
+                Kommunicate.kmConversationViewConfiguration.startNewConversationButtonTextColor = UIColor(hexString: startNewConversationButtonTextColor)
+            }
             
         } catch let error as NSError {
             print("Failed to read setting json string \(error.description)")